更新 Member 實體以使用 EAGER 取回 listings，移除不必要的 getMemberWithListings 方法，並新增 ListingServiceTest 和 MemberControllerTest 測試類別
