+++ camera.market/src/main/java/jiamingla/first/camera/market/entity/Member.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -31,7 +32,6 @@ public class Member {
     @Column(unique = true)
     private String email;
 
-    @JsonIgnore //避免無限遞迴
-    @OneToMany(mappedBy = "member")
-    private List<Listing> listings;
+    @OneToMany(mappedBy = "member", fetch = FetchType.EAGER)
+    private List<Listing> listings = new ArrayList<>();
 }
